minor punch up
